// rest/client/src/utils/types.ts
export interface DeckSliceType {
    decks: Array<DeckData>;
    status: Status;
    error: ErrorFromServer;
}
// ...
export interface LanguageData {
    languageId: number;
    languageName: string;
}

//Deck types

export interface DeckData {
    deckId: number | null;
    deckName: string | null;
    createdAt: Date | null;
    updatedAt: Date | null;
    language: LanguageData | null;
    cards: CardData[] | null;
    folder: FolderData | null;
    amountOfCards?: number | null;
}

export interface ErrorDelete extends Object {
    message: string | SerializedError | null;
    deckId: number;
}

//Folder types

export interface FolderData {
    folderId: number | null;
    folderName: string | null;
    createdAt: Date | null;
    updatedAt: Date | null;
    language: LanguageData | null;
    decks: DeckData[] | null;
}

export interface ErrorDeleteFolder extends Object {
    message: string | SerializedError | null;
    folderId: number;
}

//Card types

export interface CardData {
    cardId: number | null;
    foreignWord: string | null;
    nativeWord: string | null;
    imageId: number | null;
    voiceId: number | null;
    foreignContext: string | null;
    nativeContext: string | null;
    proficiency: Proficiency | null;
    reviewDate: Date | null;
    createdAt: Date | null;
    updatedAt: Date | null;
    deck: DeckData | null;
}
// ...
export type Status = 'idle' | 'succeeded' | 'failed' | 'loading';

export type Proficiency = 'fail' | 'repeat' | '1d' | '3d' | '7d' | '21d' | '31d' | '90d' | 'learned';

export interface ErrorFromServer extends Object {
    type:
        'login' | 'register' | 'confirmEmail' | 'update' | 'deleteUser'
        | 'deleteDeck' | 'notCreateDeck' | 'createDeck' | 'editDeck' | 'loadDecks'
        | 'loadCards' | 'createCard' | 'failedCreateCard' | 'editCard'
        | 'loadFolders' | 'createFolder' | 'failedCreateFolder' | 'editFolder' | 'deleteFolder' | 'failedAddDeckToFolder'
        | 'goal' | 'failGoal' | null,
    message: string | SerializedError | null
}

export interface FieldProps<V = any> {
    field: FieldInputProps<V>;
    form: FormikProps<V>;
    meta: FieldMetaProps<V>;
}